First commit
This change includes the first working version of memo. Memo helps you
remember short snippets of text on the command line. The program is
pretty simple. Given arguments, it saves the sequence of arguments it
received in a temporary file. Called without arguments, it spits back
the last set of arguments it was fed.

Here's an example:

  $ memo foo bar

Will write "foo bar" to /tmp/.memo. Calling memo thereafter reports back
the string of arguments.

  $ memo
    foo bar

That's it!