Refactor: replace ioutil; remove log
The ioutil package was deprecated as of Go 1.16. This commit replaces
all calls to functions declared in that package with recommended
equivalents.

This commit also removes calls to log. Instead, we simply treat error
conditions as program failures and exit with a general error signal that
other processes may respond to.