Feat: Support invocation in pipes
On nix like systems, the file mode for stdin differs when the user
invokes a program directly in comparison to when they invoked the same
program as part of a pipe. This commit adds routines to detect which
scenario we're operating under, and adjusts behaviors and argument
expectations accordingly.

Most importantly, we take the value to be 'memorized' off of piped stdin
input when the program is invoked as part of a pipeline, and we exit if
the user accidentally adds an additional argument. When we are not part
of a pipeline, we expect an argument, and otherwise print the memorized
value.