Append newlines when absent from input
When invoked directly, this command will not include a trailing newline,
since the terminal emulator uses that character as an end of input
signal. This is not desirable for a vast majority of applications, so we
append the newline in this case. In addition, refactors the signature of
write such that we only do this for direct invocations. When input is
piped, it is likely already newline terminated.